requirements.txt
fastapi
uvicorn[standard]
python-dotenv
rich
python-jose[cryptography]
sqlalchemy
psycopg2-binary
pydantic-settings